Fix workerInitFunction.
