+++ packages/engine/src/GameEngine/UserAIRuner/WorkerSupportTypes.ts
@@ -1,0 +1,5 @@
+//When adding constants update the function GetConstants() adn getWorkerInitFunction()
+export const InitRequestType = 'InitRequest'
+export const InitAnswerType = 'InitAnswer'
+export const TurnRequestType = 'TurnRequest'
+export const TurnAnswerType = 'TurnAnswer'
